python/its-status: move the loop into the module
