+++ python/its-status/its_status/main.py
@@ -6,70 +6,9 @@
 import argparse
 import configparser
 import its_status
-import linuxfd
-import sys
-import time
 import traceback
 
 CFG = "/etc/its-status/its-status.cfg"
-
-
-def loop(freq, collect_ts=False):
-    # We never close it, because we always need it; when we exit, the
-    # kernel will close it for us. That's not quite clean, but it is
-    # so much easier rather than enclosing the whole loop in a big
-    # try-except.
-    timer = linuxfd.timerfd(closeOnExec=True)
-    timer.settime(value=1.0 / freq, interval=1.0 / freq)
-
-    tick = 0
-    while True:
-        try:
-            evt = timer.read()
-        except InterruptedError:
-            # Someone sent a signal to this thread...
-            continue
-
-        now = time.time()
-        tick += evt
-
-        status = dict()
-        if evt > 1:
-            status["errors"] = {
-                "timestamp": now,
-                "tick": tick,
-                "missed_ticks": evt - 1,
-            }
-            print(
-                f"tick: resuming #{tick} after {evt-1} missed ticks",
-                file=sys.stderr,
-            )
-
-        if collect_ts:
-            status["collect"] = {"start": now}
-        for c in its_status.plugins["collectors"]:
-            if collect_ts:
-                status["collect"][c] = {"start": time.time()}
-            its_status.plugins["collectors"][c].capture()
-            if collect_ts:
-                status["collect"][c]["duration"] = (
-                    time.time() - status["collect"][c]["start"]
-                )
-
-        for c in its_status.plugins["collectors"]:
-            s = its_status.plugins["collectors"][c].collect()
-            if c == "static":
-                status.update(s)
-            else:
-                status[c] = s
-
-        if collect_ts:
-            status["collect"]["duration"] = time.time() - status["collect"]["start"]
-
-        # Here, we'd send them to MQTT or anywhere else
-        status["timestamp"] = time.time()
-        for e in its_status.plugins["emitters"]:
-            its_status.plugins["emitters"][e].emit(status)
 
 
 def main():
@@ -91,7 +30,7 @@ def main():
     freq = cfg.getfloat("generic", "frequency", fallback=1.0)
 
     try:
-        loop(freq, collect_ts)
+        its_status.loop(freq, collect_ts)
     except KeyboardInterrupt:
         pass
     except Exception as e:
